Extend program error presentations
This patch adds a Context error type that is used to form
a simple chain wherever possible with textual descriptions
of what went wrong.

This extends the current error handling story to retain
and emit lower-level errors (including those emitted from
the SEV platform itself) as it may be useful for debugging.

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>