ok: Show if Page flush MSR is enabled/disabled
Page flush MSR can be enabled/disabled. Therefore, if the flag is disabled, it
doesn't necessarily mean that Page flush MSR *isn't* supported, but rather that
it is supported yet currently disabled. So instead of returning TestState::Fail
(indicating that Page flush MSR isn't supported), return TestState::Pass and
indicate to the caller whether it is enabled or disabled.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>