Change `measurement build` output to measurement+nonce
The LAUNCH_MEASURE firmware call returns measurement+nonce. Arguably
this is two values, but qemu and libvirt treat this blob as opaque
and return it straight to the user.

Because of this, it's kind of ambiguous what `measurement` should refer
to here: the entire blob, or the value with the nonce removed.
`sevtool calc_measurement` returns the latter, but I expect most tools
will want to use qemu and/or libvirt.

Change the measurement output to return the measurement+nonce format.
If this is the wrong choice, we can add an --output-format option later.

Signed-off-by: Cole Robinson <crobinso@redhat.com>